requirements.txt
Django==2.0.2
django-admin==1.1.1
django-cors-headers==2.1.0
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.7.7
PyJWT==1.5.3
pytz==2018.3
screen==1.0.1
xlwt==1.3.0
